requirements.txt
alembic==1.0.5
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
Click==7.0
cryptography==2.4.2
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Login==0.4.1
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
mkl-fft==1.0.6
mkl-random==1.0.2
numpy==1.15.4
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.3
six==1.12.0
SQLAlchemy==1.2.15
Werkzeug==0.14.1
wincertstore==0.2
